+++ addOns/reports/CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - Update minimum ZAP version to 2.11.1.
 - Dependency updates.
+- When the automation Job is edited via UI Dialog then the status will be set to Not started
 
 ## [0.10.0] - 2021-12-06
 ### Changed